# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Chamados",
        "description": "Documentação da API de Chamados",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/setores_atendentes": {
            "get": {
                "tags": [
                    "Setores Atendentes"
                ],
                "summary": "Lista todos os setores atendentes",
                "operationId": "02e4af9279435e49b0863f0c953cfb00",
                "responses": {
                    "200": {
                        "description": "Lista de setores atendentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SetorAtendente"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Setores Atendentes"
                ],
                "summary": "Cria um novo setor atendente",
                "operationId": "600da9c3ade6016ece671251c5be144c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "setor_id",
                                    "atendente_id",
                                    "is_gestor"
                                ],
                                "properties": {
                                    "setor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "atendente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_gestor": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Setor atendente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SetorAtendente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/setores_atendentes/{id}": {
            "get": {
                "tags": [
                    "Setores Atendentes"
                ],
                "summary": "Mostra um setor atendente específico",
                "operationId": "fa45443a0babd2801fb0680cc040871b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do setor atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do setor atendente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SetorAtendente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Setor atendente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Setores Atendentes"
                ],
                "summary": "Atualiza um setor atendente existente",
                "operationId": "4c5136b610ca01a6dcaa5c4189766e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do setor atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "setor_id",
                                    "atendente_id",
                                    "is_gestor"
                                ],
                                "properties": {
                                    "setor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "atendente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "is_gestor": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Setor atendente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SetorAtendente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Setor atendente não encontrado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Setores Atendentes"
                ],
                "summary": "Deleta um setor atendente",
                "operationId": "473410374b4cdf502f38027db5382e51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do setor atendente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Setor atendente deletado com sucesso"
                    },
                    "404": {
                        "description": "Setor atendente não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SetorAtendente": {
                "required": [
                    "setor_id",
                    "atendente_id",
                    "is_gestor"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "setor_id": {
                        "type": "integer"
                    },
                    "atendente_id": {
                        "type": "integer"
                    },
                    "is_gestor": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}